+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- accept both tag and version as input
+
 ## [1.3.0] - 2021-10-12
 
 ### Added